Split Geospatial.hh
